pyproject.toml
[project]
name = "quant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair>=5.5.0",
    "asyncio>=3.4.3",
    "cupy-cuda12x>=13.5.1",
    "ipykernel>=6.29.5",
    "japanize-matplotlib>=1.1.3",
    "jax[cuda12]>=0.5.3",
    "jupyter>=1.1.1",
    "numpy>=2.3.1",
    "plotly>=6.2.0",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "torch",
    "yfinance>=0.2.65",
]

[[tool.uv.index]]
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/nightly/cu129"
